Split huge class into separate classes
PHPStan already complained about huge complexity.
Split type building and adding of properties to dedicated classes.
